:rewind: [Vitest] Move test folder back to original location
