+++ src/useEventListener/tests/useEventListener.test.tsx
@@ -3,7 +3,7 @@ import {screen} from '@testing-library/react';
 import {renderHook} from '@testing-library/react-hooks';
 import userEvent from '@testing-library/user-event';
 
-import {mount} from '../../../test/utilities';
+import {mount} from '../../test/utilities';
 import {useEventListener} from '../useEventListener';
 import {AllowedEvent, EventListenerComponent} from './EventListenerComponent';
